#2712 refactored tabs component
