+++ src/frontend/src/hooks/organizations.hooks.ts
@@ -9,11 +9,6 @@ interface OrganizationProvider {
   selectOrganization: (organizationId: string) => void;
 }
 
-/**
- * Custom React Hook to get an organization.
- *
- * @param organizationId organizationId to be fetched.
- */
 export const useCurrentOrganization = () => {
   return useQuery<Organization, Error>(['organizations'], async () => {
     const { data } = await getCurrentOrganization();
