+++ src/frontend/src/pages/AdminToolsPage/AdminToolsPage.tsx
@@ -14,7 +14,7 @@ import TeamsTools from './TeamConfig/TeamsTools';
 import AdminToolsBOMConfig from './AdminToolsBOMConfig';
 import AdminToolsProjectsConfig from './AdminToolsProjectsConfig';
 import { useState } from 'react';
-import NERTabs from '../../components/NERTabs';
+import FullPageTabs from '../../components/FullPageTabs';
 import { routes } from '../../utils/routes';
 import { Box } from '@mui/system';
 import AdminToolsRecruitmentConfig from './RecruitmentConfig/AdminToolsRecruitmentConfig';
@@ -72,7 +72,7 @@ const AdminToolsPage: React.FC = () => {
       title="Admin Tools"
       tabs={
         <Box borderBottom={1} borderColor={'divider'} width={'100%'}>
-          <NERTabs
+          <FullPageTabs
             noUnderline
             setTab={setTabIndex}
             tabsLabels={tabs}
